Adding --cordova flag for "recroom new"
Passing in the --cordova flag creates the cordova app structure.
We create the dist-cordova directory, which contains the Cordova xml,
hooks,platform and www directory.

"cordova create --link-to dist dist-cordova..." symlinks
dist-cordova/www to dist.

TODO
Need to add a check to see if cordova is installed and warn the
user if it is not.